--- a/index.js
+++ b/index.js
@@ -7,4 +7,17 @@
  * Licensed under the MIT license.
  *   https://github.com/estools/espurify/blob/master/MIT-LICENSE.txt
  */
-module.exports = require('./lib/clone-ast');
+'use strict';
+
+var createWhitelist = require('./lib/create-whitelist');
+var cloneWithWhitelist = require('./lib/clone-ast');
+
+function createCloneFunction (options) {
+    return cloneWithWhitelist(createWhitelist(options));
+}
+
+var espurify = createCloneFunction();
+espurify.clone = function (ast, options) {
+    return createCloneFunction(options)(ast);
+};
+module.exports = espurify;