--- a/index.js
+++ b/index.js
@@ -17,8 +17,5 @@ function createCloneFunction (options) {
 }
 
 var espurify = createCloneFunction();
-espurify.clone = function (ast, options) {
-    return createCloneFunction(options)(ast);
-};
 espurify.customize = createCloneFunction;
 module.exports = espurify;