--- a/package.json
+++ b/package.json
@@ -9,8 +9,7 @@
   },
   "bugs": "https://github.com/estools/espurify/issues",
   "dependencies": {
-    "indexof": "0.0.1",
-    "traverse": "~0.6.6"
+    "isarray": "0.0.1"
   },
   "devDependencies": {
     "esprima": "^2.1.0",