… lib/ast-deepcopy.js