Optimize AST clone function

- [x] optimize
- [x] make backward-compatible
- [x] provide customization function
- [x] documents
- [x] benchmark


```
$ node bench_babel_tree.js
espurify v1.1 x 584 ops/sec ±0.83% (91 runs sampled)
espurify v1.2 x 8,022 ops/sec ±3.81% (87 runs sampled)
npm clone module x 56.54 ops/sec ±2.10% (59 runs sampled)
Fastest is espurify v1.2
$
```

```
$ node bench_deep_tree.js
espurify v1.1 x 10.53 ops/sec ±8.56% (30 runs sampled)
espurify v1.2 x 196 ops/sec ±2.35% (77 runs sampled)
npm clone module x 1.56 ops/sec ±4.23% (8 runs sampled)
Fastest is espurify v1.2
$
```
Benchmark is powered by [Benchmark.js](http://benchmarkjs.com/)